New Gas Mask
If you want to get the holes to work you’ll have to import the
textures, create a material out of them, go into the material editor,
and connect the bottom white circle on the texture sample to the
opacity mask. Then in the details on the left go to the material and
change the blend mode to masked.